storybook: add warning story
