+++ frontend/packages/core/src/Feedback/index.tsx
@@ -1,4 +1,5 @@
 import Hint from "./hint";
 import { Note, NoteConfig, NotePanel } from "./note";
+import Warning from "./warning";
 
-export { Hint, Note, NoteConfig, NotePanel };
+export { Hint, Note, NoteConfig, NotePanel, Warning };
