+++ frontend/packages/core/src/error.tsx
@@ -5,7 +5,6 @@ import {
   Collapse as MuiCollapse,
   ExpansionPanel,
   IconButton,
-  Snackbar,
   Typography,
 } from "@material-ui/core";
 import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
@@ -90,34 +89,4 @@ const CompressedError = ({ title, message }) => {
   );
 };
 
-interface WarningProps {
-  message: any;
-  onClose?: () => void;
-}
-
-const Warning: React.FC<WarningProps> = ({ message, onClose }) => {
-  const [open, setOpen] = React.useState(true);
-
-  const onDismiss = () => {
-    if (onClose !== undefined) {
-      onClose();
-    }
-    setOpen(false);
-  };
-
-  return (
-    <Snackbar
-      open={open}
-      autoHideDuration={6000}
-      onExit={onDismiss}
-      onClose={() => setOpen(false)}
-      anchorOrigin={{ vertical: "bottom", horizontal: "right" }}
-    >
-      <Alert elevation={6} variant="filled" onClose={onDismiss} severity="warning">
-        {message}
-      </Alert>
-    </Snackbar>
-  );
-};
-
-export { CompressedError, Error, Warning };
+export { CompressedError, Error };
